requirements.txt
aiofiles
aiohttp[speedups]
beautifulsoup4
cloudscraper
dnspython
ffmpeg-python
moviepy
opencv-python-headless
pymongo
pyrogram
pyromod
python-decouple
Pillow
PyYAML
TgCrypto
yt-dlp
wget
